requirements.txt
asgiref==3.2.7
cffi==1.14.0
cryptography==2.9
Django==3.0.5
django-rest-knox==4.1.0
djangorestframework==3.11.0
Pillow==7.1.1
pycparser==2.20
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
